Update issue 2044
Shuffling some of the logic so that it is compressed more efficiently. I had a
couple different times when I was checking to see if the AOI was provided to do
certain operations. In this case I moved some distance handling logic into the
original check to see if this information was provided.

Reduces uneccesary code.


Former-commit-id: b7c818d9a9524fe44163a1b96eb5ee684ab3aa17