Update issue 2202
Added a secondary loop to my zeros_check function in hra_preproc. Now, in additon to pulling any pairings where there are no criteria for h_s_c nor h_s_e, we are also checking for instances where either E or C are represented, but not both. If there are 0's in all criteria for a pair (which would have been removed in the for loop above the new one) then we will now remove the whole pair from both h_s_c and h_s_e, which will keep it from being used in the assessment. Addititionally, if there is an E without a C (h_s_c[pair] contains something, and h_s_e[pair] does not, or visa versa) we will throw a MissingEorCError, and inform the user to change the criteria spread within the appropriate pairing on the appropriate ratings csv.


Former-commit-id: d0230bac03832e3ed2e8f87a85fea83c06096dab