Update issue 2284
Adding in logic for Option 2 described above. If a baseline threat raster can
not be located, the model will not exit, but will not try and calculate Habitat
Quality for the baseline land cover. Instead it will skip to rarity.

If a threat raster can not be found for cur or fut land cover, the model will
exit gracefully.

I tested a few of these conditions and it seems to behave properly.

Should test more on my work machine


Former-commit-id: f846cef82a2176b87e63ae5889caf82a05327275