Update issue 1573
The zero_check no longer in charge of the exception raising for 0 dq/w. Want to make sure that if we are erroring for a 0 in either of those two, that it's not being placed there because someone is trying to remove a criteria. So, adding an additional condition to the 0 check statement that says if rating is 0, leave it in. That way, should get to the zero_check function later and be removed from there after the fact.

Had I thought this out better ahead of time, it wouldn't have been this messy, and could have been all one function.


Former-commit-id: 71acb0a2a89fd377329ba754a3f4b02251d50aa6