Update issue 2301
Making a change to aggregate raster values so that rasters and data sources are not required to be
projected.  This is helpful if we have global datasets with just lat/lng values and we don't care
about getting the projected units right as much as we care about min/max pixel counts.


Former-commit-id: dc465b5728e8f3154c1f91f4d4350c3f601b65fe