Update issue 2367
For AOI plots, was using the upper bound on a plot to be the max score that a user provided, multiplied against the number of stressors. However, since each AOI has their subplots on a habitat basis, then subplots by stressor within that habitat. Thus, the highest value on a graph should only ever be the max value for a h-s, not highest value for a habitat.

Changed all instances of "max_value" to just the the args item max_value that the user is providing.


Former-commit-id: b5bc61d60bec852216717cdc4a7b91169638680f