Update issue 2365
Starting on new functionality for the zero_check. Determined that will be left in hra_preproc, but it will pass a variable back to hra within the massive dictionary. Like the others, it will get passed on to hra_core directly.

This variable will be a 'warnings' dictionary containing two items. 'print' will map to a list that contains any statements that I desire to be printed using a logger.warn() right at the end of a run. 'unbuff' will contain a list of pairs which do not have e criteria. For these, we will skip the concatenation of all E numerator files (there won't be any) and use the unbuffered stressor raster instead. This will provide spatial overlap in lieu of exposure ratings.


Former-commit-id: 71df2d6dbd1b8c0ab4b2215dc801f928f9418426