Update issue 2365
Fixing the calc_E_raster call to now only happen if the pair we're looking at isn't in the warnings['unbuff'] subdictionary. We will call a copy_raster function that will be created in order to put a copy of the unbuffered raster (from the 'Stressor_Rasters' file) into the E_Risk_Raster.tif file. (Since, by convention, anything being used as E should really be in there.


Former-commit-id: b906b526c93838f8a1d8643ca7db741e2b484585