Update issue 2365
Adding if statements to the pre_calc function so that a denom value only gets added if there is data within the corresponding h_s_x dictionary. Do not want to be initializing them to 0 if that's going to be the only value, since it will cause div by 0 errors down the line.


Former-commit-id: 8ddde26d31f3661e533125fb5027857472385191