Update issue 2365
Instead of pulling denoms from the dictionary if they're 0, we're going to allow them. The only circumstance in which we should get denom values of 0 is if as criteria exists for habitat, but is 0. Everything else would have been taken out. Thus, we're going to allow denoms to contain the 0 value, but short circuit with an if statement to check for 0 values before we would need to divide by it.


Former-commit-id: 316c0b26cae1600719a103c001897ec833c220ed