Update issue 2365
Changing C raster calculation so that if a 0 comes in (habitat exists, but has no criteria score), it can be returned as 0. Additionally, in the risk calculation, we had some short circuit logic in which the hab data on its own was being returned as 0. Instead, if we know that hab has data, but that there is no overlap, want to retun whatever the c_pixel is, NOT 0.


Former-commit-id: 903b4421ca5f84798230b09c6b032060e304c948