Update issue 2360
Making some changes to the layout of table_generator.

I think I am going the way of building up strings for each element to be added
to final html report. That is, the create table funciton will return a string
representation of the table in html format which will then be added to a
dictionary which will then be constructed into an html page once everything
that is wanted on the page is added.

I think I might be mis using this issue here, will have to layout a better
structure for issues.


Former-commit-id: 92dc7c036d8c58e7d400082d8e4f427209d0f2c1