Update issue 2362
Adding the param max_risk to the pre_calculation call. We will use it when calculating the spatial overlap scorebased on pixel overlap.

Also created aggdictionaries for habitat raster and h_s raster which will return the pixel total of each within the given subregion. We will then divide those to get the h-s fraction of overlap, and use that to get a spatial overlap score that can be averaged in with our calculated E score.


Former-commit-id: 75edf82ead890e9b3806821a54a5cad5b7d78a27