Update issue 2384
This also touches issue 2360.

I am rearranging how the reporting API is structured. There is now a reporting
package that currently has an __init__ file as well as a table_generator module.

I have started a lot of work on setting up the structure for generating tables.
Currently I am going back and forth on the best way to create a table given
different inputs. For examples someone should be able to create a table from a
group of polygons (shapefile, and how permitting will work), a CSV file (a lot
of our models output CSV files and transporting it to html should be quick), a
dictionary (sometimes we just build up dictionaries and should be able to port
that to a html table).

Ultimately, the function that will generate the html table expects a dictionary
in a certain format. Thus pre-processing steps / functions need to be made given
the other input types.


Former-commit-id: 6b6c5ad92ac1a4d6c9f410e73d37f4d644ea42c5