Update issue 2393
Updating 'build_table' which is a function that handles table elements passed
in to the report generator. Updating this based on the new structure I'm
working on for passing in a shapefile/csv/dictionary for the source of the
table data. Depending on which type of input is given, it structures the data
and options accordingly to pass on to the generate_table function


Former-commit-id: 333a0ef1e1adb0afef86763642f70d5bafe52291