Update issue 2412
Adding a new assertion to the testing package.  This new assertion loops
through two text files line by line and asserts the equality of the
strings representing each line.  If an assertion is found, a helpful
error message is provided.

This functionality is tested and will be used in the AssertFiles
function in GISTest for files with .txt and .html extensions.


Former-commit-id: 362ba99469c67f4839c380cc05321a0ce9d79e56