Update issue 2421
I decided that it was best to change the structure of how we were passing in
column definitions. Before I had a dictionary with the keys being column names
and a value inside the sub dictionary being an id. I think it makes more sense
to have the keys be the id's and have the column name nested in the sub
dictionary. This makes current code shorter, asserts we won't have a duplicate
column position, and helps with adding checkbox columns.

Added a skeleton function for adding a checkbox column

Changed a lot of doc strings to update the change in column definition


Former-commit-id: 59cde9b9baa8606d2625a376c1aa232726435c4f