Update issue 2421
Adding logic in for adding a checkbox. Now a 'checkbox':True/False parameter can
be specified in the table element. If True this will trigger a function in the
table_generator that inserts a checkbox column as the second column in the
table.

Added the new checkbox function and doc strings / comments


Former-commit-id: 0c3b0bcfb7522d037f57623664614ccf276ba78c