Update issue 1016
Making the lastrun functionality work as expected for dropdowns that are
affected by tables input as another element.  Involves remembering a
state that is a little more complex than what other elements do, but
hey, it seems to work!

The complexity here is that we only activate the old state so long as:
  * a state has been loaded, whether from lastrun or a parameter file
  * the current table file for the dropdown matches the table URI that
    affects this dropdown
  * the key at the saved index matches the key selected when the state
    was saved.


Former-commit-id: bde99c73730adb509e5f2e6b18bd44cb478eb89e