Update issue 2424
Making a big update to how totals are updated. This update ensures that only
columns where the user has requested be totaled, are totaled. This great
functionality will put the responsibilty on the user to try and not 'total' text
or other non totalable fields.

This commit also has a lot of tightening / cleaning up of code and uncessary
conditionals. Added a function 'get_column_total_list' that returns a list of
booleans that correspond to a list of column names. These booleans indicate
whether the column should be totaled.

Modified 'add_total_row' to be smarter about how it built up the table data for
hte total row


Former-commit-id: f265324d88431ca5940a91c68bf32e46e64764fc