Update issue
I ran into a situation where I was making specific funtions for certain tasks
when each task did pretty much the same thing. I thus decided to remove the
duplicate code and make a single function that defines the task it does and
does not label it as being specific to something.

I removed get_column_headers and get_total_column_list and replaced with a
function that takes a dictionary and key name and returns a list of inner
values in order of the main dictionaries keys.


Former-commit-id: e9d4914e822847a470985b304787457fc709810f