Update issue 2434
Updating sub repo because of jQuery changes and added test.

I updated my jQuery functions for totalling to handle multi table elements.
Before it was specifically looking for table id's, however with multiple tables
it needed to be more general. So now for constant totals I loop over the table
elements performing the task. For check box totals I get the local table from
where the event.change is coming from and isolate all my moves to that scope as
to not conflict with the other table.

Looks good!


Former-commit-id: 13b97984446a7093a5aceb056cd0f4631e653c27