Update issue 2405
Beginning restructure of how my html_obj (dictionary) is built. In order to
ensure a reasonable order of element positions if there is a duplicate position
I have now saving the elements as a list of strings with a list of positions
that I then am able to set the proper order later.

If a duplicate occurs, they could end up in before or after the other


Former-commit-id: 8d04a043558e6fdb56175edad1f096a3d0f72de5