Update issue 2405
In order to ensure that a duplicate position in a section didn't blow that
element away I had to modify the structure of how I was storing the data in the
html_dictionary.

I decided to have the keys 'head' and 'body' point to a tuple of lists where the
first list was of the string elements and the second was of the corresponding
position. This way when I went to write out the elements in the html page I
could properly order duplicate positions.

Had to update a lot of comments and docstrings


Former-commit-id: 95e6af3955fc85e28ee484a6bb96a64288cdcb91