Update issue 1230
Updating the way that C is being calculated. Want to first average the resilience criteria so that they are treated as a single unit. Then that gets added to the sensitivity criteria, and summed, all of which gets divided by the summed denominator (which, in turn is the averaged resil. + the rest of the sensitivity denoms).


Former-commit-id: 19b2859ed1277ff2b0a21b80da78b77aae9ba1a4