Update issue 1955
Editing how soil storage is calculated. Basically we are now calculating it on a watershed basis. Wrote a new function that creates a new raster from the equation minus the inclusion of streamflow. Then aggregate that new raster to get the means. Once we have the means we get the equivalent volume and subtract the streamflow volume from it for that watershed.


Former-commit-id: 418613712dab38c3ced484ccf5bb8aef3c95ec61