Update issue 2484
Making a lot of white space fixes from programming on Windows. This is what
broke the build by the way.

Also making a lot of bug fixes from all of the new functionality I added.

Changing the output columns of the CSV to be Streamflow_vol, Streamflow_mn, and
Storage_mn. Also computing these values.

Model runs and generates values


Former-commit-id: a507398a74ca03655b9d145a3636dd2e13a44f2c