Update issue 2484
Fixing what the streamflow mean is reported as. Before I was adding a sum with a
single pixels mean. This gave high false values. Now I take the max dflow,
divide by number of pixels to get a dflow mean and then add. I'm still not sold
this value makes sense or is helpful, but i'll keep it in there


Former-commit-id: 476c4840f6faf557effce298249077a19951060d