Update issue 1230
Changing the E calculation equation to match the C in that there should be habitat shown where there is no overlap but habitat exists. Adding h_base_uri as an input to the E calculation equation. Contains location of the desired habitat raster. In those areas where there is habitat, but not overlap, we will return 0.


Former-commit-id: 935f45334967c3e26fd3d1ddf34a7c02e6b77d45