Update issue 2424
Updating how add_totals_row function deals with row / row data class names. We have specific names that jQuery / javascript function looks for and I don't think at this point it makes sense to add flexibility for the user / developer to name their own classes here. Thus I am replacing those in the signature with a boolean value to determine if the total row being added is a checked box total row or a regular total row. An IF statement then sets the class names for the row / row data


Former-commit-id: b92fe14037f186806d87f25081bc6f95cd7cbde3