Update issue 2484
Setting imperv fract to 0.0 if Stream is present. Trying this becuase there is a bug in absorption equation, where by setting streams to have imperv fract of 1.0, absortption is 1. This is wrong, if there is a stream then absorption should be 0.0.

I'm not saying this is right, but I wanted to see how the outputs would be affected. They do look better, yet the water volume balance still doesn't look quite right.


Former-commit-id: f558af4f3fd7eeb2fdab88f4f4aa93323a211adb