Update issue 2529
Adding alternative AET equation.

Reading LULC_veg column into dictionary. Creating a raster from that file and using that raster of 1s and 0s as a mask for which equation to use in 'fractp_op' in cython and in aet_op.

ALSO, removing the divide by 1000 for Kc values. Little double dipping here.


Former-commit-id: 2b176c10ed8868693d0d61e3a5fab8139886661f