Update issue 2484
Updating the in_source value for direct flow to be precip * cell_size ^ 2.

Rich seems to think that is what in_source should be.

I attempted to set impervious layer to all 1.0 and expected dflow to be the total precip. Instead dflow came out 0s. I believe there is a bug in routing_utils.routing_cython_core. I think either the equation is wrong for flux_and_source or we are thinking about absorption backwards.

Currently in routing_cython_core, if 'source_and_flux' (line 168) flux gets set to flux * absorption. So if our absorption is 0s (cause imperv layer is all 1s) then flux is 0s. This causes issue on line 204 and 209 where we set flux to an equation that is multiplied by 0. Thus we get 0.


Former-commit-id: 1618ff043e504d8db99f036565e6824a62bf51f2