Update issue 2484
Quite a big update. I noticed a problem in routing on a small sample dataset that slope wasn't being computed for the edge pixels which in turn caused any water on those edge pixels to not be routed. I believe this is what is throwing off the water balance.

I thus attempted to set up the code using a buffered DEM, not the clipped or aligned one, so that the slope values would be calculated. This is very much set up for the sample dataset and not real data. I was able to get full routing using these methods, and also make a slight adjustment to routing_utils.route_flux.

I reverted route_flux so as not to mess any other code up.

I think there are a couple options to this issue. One is to not take into account the edge pixels and mask outputs by the created slope raster. This way at least we can know if we are getting a correct water balance. We could also attempt to buffer the DEM in some way so that slope values were calculated for the edge pixels.


Former-commit-id: 122d3b91f4294ee2d696dba6585decac8c86d261