Update issue 2564
This commit reflects the attempt to go with option 2 from above; don't include the edge pixels in calculations. What this means is we mask out the precip / evap rasters by the slope raster such that only pixels with a slope value are accounted for in the water balance.

I swapped around some functionality, clipping the dem first thing, calculating slope from that, then aligning all the input datasets.

When creating the precip / evap rasters I now run a vectorize_datasets function that basically adds a slope nodata mask.

Doing this on my sample dataset gives me a water balance of 0.

Next I'm going to try it on the larger datasets


Former-commit-id: d455d80b0dba9cae8dce6a4bb915a89a205c3ce0