Update issue 2407
Figuring out an error I get from trying to label the svg using the ws_id attribute. I found out I was getting the error because Kartograph expects the label attribute to be a String.

I think in this case if we want labeling we'll have to run a preprocess step where we copy the shapefile, convert all fields to strings.


Former-commit-id: 7bafd8d4ebd83b6984f650e40a47036cea429ced