Update issue 2493
My first attempt to make script paths relative had many bugs, fixing them here.

Updating the tests to use proper location of script files which are saved in
invest-data/test/data/reporting_data

The function for adding head elements and dealing with scripts was slightly off.
I copying the file over to the proper location, that of the output html file,
however I was setting the script path in the html page as a full path not
relative path. So it was like: 'invest-data/test/data/reporting_data/script.js'
now it is: './script.js'.


Former-commit-id: 7e396ef49f081258fc104ea39312bf6d8f9799d7