Update issue 2494
Taking a stab at making inputs for reporting lists instead of dictionary. The
first thing I've changed is making column and data inputs be lists of
dictionaries. The user then is responsible for properly ordering the
dictionaries in the list as they want it represented.

I modified one test and gave it a run, everything looks like it functions
alright. Will have to change the other tests to conform to this new setup.

Also will need to remove commented out code and update comments / docstrings


Former-commit-id: f36daffa0893235b0e93646bc52d70e19be7b238