Update issue 2494
Removing the position attribute from being needed in the elements dictionaries.
I was using position to order 'body' and 'head' html elements. However, I think
it better that they just appear as ordered in the list in which they are passed
in. This reduces complexity and code and puts a very slight owness on the user.

Hopefully this will make the API lighter, more user friendly


Former-commit-id: e469b3c515e6f686831f8b2aa58863140ba6316a