Update issue 2564
Going through MWY and cleaning things up. I am no longer masking the precip / evap rasters to disclude the edge pixels as I believe Rich is taking care of this in routing.

Commenting out code that is not being used as we now are assuming interflow / baseflow are routed, thus we are only interested in the max values


Former-commit-id: 44e616ef3345128f3b29c9d2e61288a442c8fb0b