Update issue 2482
Major overhaul of including sub watershed outputs and calculations. Basically copied the aggregation for watershed and resulting outputs and did the same over the sub watershed.

Made a change in how values were aggregated over. There was a lot of duplicate code so I condenced it into a for loop that zipped over a couple lists. This now returns a dictionary of the max pixel values for each raster.

I'm sure things will be broken. Look for bug fixes shortly.


Former-commit-id: 385b0f817175627c4b8264f3b9b32a47ec3217a3