Update issue 2127
Apparently I have saving these changes in the wrong issue, issue 2482.

I have been making many overhauls / modifications for implementing sub watershed functionality. I reorganized some of the aggregating code that go the max pixel value from various rasters. I then took the watershed section and copied it and applied it to sub watershed calculations. There may be a way to not have so much duplicate code here.

This commit was supposed to be a fix for bugs that were introduced from the above changes. Ran the model with sub watershed file and it doesn't crash!


Former-commit-id: ebe251d2929d6f35e60319f148e89190a68a9a7a