Update issue 2554
For the initialize_pop function, ended up only using this once, for the first cycle. So removing the need to pass in the cycle number param, since we know it will always just be 0. Just manually set it to 0 as a dictionary at the beginning of the function, then explicit call 0 throughout the function.


Former-commit-id: 0ecac62859ea1f0ee9bc13eddb1061c6726841d0