Update issue 2484
Wrote a function for calculating streamflow at the pixel level. I don't necessarily need this for updating the storage issue and am not sure if it will be needed at all. I think in the end we want to figure out where the outlet pixel is on the landscape within each watershed and combine those 'maximum' values. That functionality will be added when Rich completes the new routing operations.

But this function can live here for now incase we end up wanting streamflow at the pixel level


Former-commit-id: a97e0a68b2506967038d04d0c7b9098c34b8aa0a