Update issue 2554
We are now assuming that if the user desires their harvest in terms of weight, that they would also want their spawners in terms of weight. Thus, we're now pulling a weight from being straight multiplied against the harvest, and adding it into the spawner counts.

THIS MEANS:
For any model run where the user wants to see their harvest in terms of weight, we are actually running the entire model with biomass instead of an individual count. SO, pulling the aftermarket multiply so we don't end up with biomass * weight when we do the harvest data. Instead, do_weight is being passed down through all the cycle equations into the recruitment equation, and is being taken into account when spawners are calculated. If weighting is desired, the number returned from that will be in terms of spawner biomass. If not, it will be a number of spawner individuals.


Former-commit-id: 814e81e72b00980ae3428b9365edfb6fee8fb8d1