Update issue 2589
Adding bulk of logic to permitting JS file for table interactivity. After some simple clicking around a sample data file, all functionality seems to be present.

There is a lot of duplicate code and cleanup should be done.

Added a fancy 'fadeOut' when removing a row, it can be removed if it doesn't look good.


Former-commit-id: 3a08e41fc0236bd46525a7cfab598a8884c8cd9f