Update issue 2617
Was able to get embedding of SVG to work. Now have an SVG type and add_svg_element function that creates an SVG file from a shapefile by calling to style.py. Then that file is read in and embedded as html in the report. Looks OK, obviously need proper spacing and buffers between elements.

So, I think we have the ability to embed SVG.


Former-commit-id: c695088fe55483d7071f7523421faeb4f87b60bf