Update issue 2554
Actually don't want to be summing across multiple years, since we wouldn't actually be "harvesting" those years. Removing the totals_dict from the cycle function. Need to remove all references to it in other functions, and replace them with the final harvest. Should perhaps return the number of cycles it actually took instead.


Former-commit-id: 3b896c21b688771f4f4ccd0f0083d6a441833f23