Update issue 2554
Fixing the subregion copy thing to reflect that we're now passing in the final cycle results to it, instead of the totals dictionary. Should be noted that we don't need to ignore the 'Cycle_Totals' here, since the keys are being generated from the features instead of the cycle keys.


Former-commit-id: a5d5faa8094522439102058593e14cdbc1bbc538