Update issue 2554
Fixing the html output to play nice without using totals dict. Do have to check within the hrv_dict subdictionary for the 'Cycle_Totals' key, and except it. Also, are only setting the 'equilibrated' Y/N thing to Y when we get to the cycle where equil_pt == the cycle number. This also should be the last cycle in the table.


Former-commit-id: 45ddf58476e2570f933566deec4bf7c07b843c53