Update issue 2650
First hack at getting classes to column headers. I added the ability for an option in the column definition (list of dictionaries) to add attribute elements as a dictionary. So: 'attr':{'class': 'offsets'}

Wrote a helper function in table_generator that creates a nice string from these attributes and adds it in the <th> tag. I think I can remove some old column sorting code in table_generator but have to look.

Added a new test for using classes

added documentation in __init__


Former-commit-id: c3603a7acb959f6225b1d20e143243ff6ad943eb