Update issue 2650
Adding ability to pass in table attributes. I added the option for passing in an 'attributes' key in the table dictionary, that points to a dictionary where key value pairs can be added for attributes.

This way we can add 'id', 'class', 'border', and others to the table so that CSS can do awesome magic with it


Former-commit-id: ea2a8fdcc77cba1ab060a1be186269c86d423ab7