Update issue 2647
Added a function in the permitting js file that looks for any html elements that have a class name called 'scientific'. If so, the function takes each instance that has class and sets it's data elements to scientific notation if the values are numeric.

Now that I write this, I suppose I should specifically check to see if any table headers or <th> tags have the class name 'scientific'


Former-commit-id: 7addc39c58a005d4d0950f3d2e884894875be8b6