Update issue 2654
Starting a function to handle the initial impacts. When doing this I realized that since the column structures has changed in the last table, it will mess up my current handling of adding parcle benefits when selected. This is because we now have extra columns in the table that are not necessarily represented in our parcel_json data file.

Need to go through and come up with a new method for handling this.

I did write a little bit of code that gets the column headers from the last table and sorts them. Hopefully this can be used to help address the above


Former-commit-id: 455a33640ff33e84fdbf6a5db5105485fcb5d94a