Update issue 2654
Replacing code for updating a muni when a parcel id is selected. This was updated only for when that muni already exists.

Moving where impacts gets initiated to the main function. That way I can build up the muniState object to include the munis that get added for impacts. This will save on having to test another case when a parcel_id updates via checkbox


Former-commit-id: 55dde690bdb065949c7af5e3e229742a8181d10f