Update issue 2679
Sat down with Perrine and we've come up with a workable solution for the negative values issue.

The new aet/p equation is OK, but we need to adjust how the omega (w) is handled, which deals with the seasonality or Z constant.

We now calcuate w by dividing the value by 0.2. This according to resources will allow us to keep our seasonality constant range (1 - 10) and will put W in it's proper scope of hopefully between 1.25 and 5.0.

However, if the input values for a users data are different or extreme, the w value could be outside these bounds which then cause the new aet/p equation to go negative. Thus, we bound w between 1.25 and 5.0, saying that if it falls outside we are going to group it as an extreme case of the upper or lower, but not let it break the form of the data/research/graphs/ model.


Former-commit-id: e5902f4f04bea55e56fe694399d23eed8c07363a