Update issue 2679
Perrine came up with better way to handle the Omega value for the AET\P equation. She grabbed this method from one of the Donohue papers. Instead of dividing by 0.2 we will now add 1.25 onto the original (awc/precip) * Z. This guarantees a minimum bound for Omega so we only need to check the upper bound at 5.0.


Former-commit-id: b512dc0cbb5124b3c68f1a731109e1b6b7064e5f