Update issue 2654
Adding a little object for when a muni row is updated by a parcel_id. Setting the 'count' key in the muniState dictionary to 1 if not present. This is for all the munis that were added for impacts and that should not be removed. Since it indicates an impact muni I set to 1 even though it has been accessed twiced. This should ensure that on removing the check boxes it stays as a row


Former-commit-id: d8046130917a0477f6df011331fcce49784eca24