Update issue 2654
Adding initial functionality for updating and adding a new municipality on a parcel_id check. Have not tested but want to commit this before moving on to handling deletion of rows and unchecking a parcel_id.


One Problem that has arisen is how to handle population column. Initially we decided to pass it in with the impact json data. However, clearly this does not include all the munis. We thus will have to find a way to add it to the parcel json data.


Former-commit-id: 9db36d0a9e06a8830d7519c0efc3b612599b0163