Update issue 2683
Adding a constant for the reporting_data folder using __file__

In order to abstract default files from user files, I removed the totalling functionality from the permitting_functions file. The totalling functionality will be separate and will be a default file that is added if 'Total' is present.

I should be able to add the defaults at no expense to the User.


Former-commit-id: 97adec92c60d02cba485e61636c6d98fb06c0101