Update issue 2450
Creating an object to manage data used/set/reset by the various
components of our multithreaded UI.  This object uses threading locks to
ensure that the data being read or written is done so in a transactional
way and in atomic chunks.  Pretty primitive, but I can see where this
approach might be very handy later on.

The second part of this commit uses this data object to manage the
location of the last folder visited.  If a File input has text in it,
then the file/folder selection dialog should open to the folder
containing that file, or else the folder selected (if a folder dlg.)

If, however, the file/folder field is empty (as commonly happens in
RIOS), then the file/folder dialog opens to the folder last visited.

This functionality does _not_ guard against whether a folder does or
does not exist before its value is returned ... I'm not sure what would
happen if the last folder selected did not exist.  For future pondering.


Former-commit-id: 0cb4b4756176f4d5b7d33aa56039facf73439ce6