Update issue 2775
Came up with a way to progegate class names to table data elements based on columns.

When defining a column, there is an optional key 'td_class' that can be add that points to a String for the class name you want to assign to the table data for that column. This will not assign that class name to the table header, but will assign it to any totalling row data elements.

Added a helper function in table_generator for creating a data structure to handle 'td_class' elements. Had to update a few functions for creating the totals row, that handled the 'td_class' elements.

Added a new simplified test to make sure the 'td_class' keys were being handled.

Added documentation into __intit__


Former-commit-id: a1886d35a275ee019c83c67aab9c299484989299