Update issue 2842
Updating vectorize_dataset functions in habitat quality.

There is one function map_ratio that calls into an outside dictionary to get the
proper value to set. I'm not sure how to take this and make it into a numpy
operation.

So there is still one vectorize_dataset call not fully converted


Former-commit-id: 702b90347d1b43b628b579f2912bec2ceb384433