Update issue 2869
The check for 0 rating value in the beginning of a line where there are also 0 DQ/W values was actually int/float instead of strings (which is how they come in from the csv). Need to do the check as a string, since we won't cast them until we're entering in the dictionary.


Former-commit-id: 919fa2d4f6070771a3107a135210024d6649c85c