Update issue 2936
Added very explicit conditionals about handing land points from the CSV file. If land points are present do business as usual. If no land points were provided, then just compute the distance from ocean to grid points as the crow flies and use that for the cable length in valuation.

Currently I have a lot of if/else statements that can be cleaned up, but I did an initial test and the three different cases seem to be working properly.


Former-commit-id: 33bb433e1b24e276203d54e9ef4ad6ce797d6a48