Update issue 1726
Rob is working on a project where it would benefit greatly from being able to have a table of wind energy prices over the lifespan of the farm.

We took this opportunity to bring some consistency in valuation to the models. We decided to replicate the Blue Carbon model and have an input for a CSV price file and if no price file, inputs for an initial price and then an input for a rate of change.

Added new inputs to the UI / json file. Wrote up help text.

Built in valuation functionality to the model for the case of using the CSV file and the case of using the price and percent change.

Have done a lot of testing, should be ok.


Former-commit-id: 24e7e01f42e9c71909f4eed9886eca17225f1d0f