Update issue 2995
Was initially testing the output with a scalar of math.ceil(num_habs/4). But still seems to be compressing the labels on the plots. Trying to stretch to a constant size now to see if I can do math.ceil(habs/4) + 1 and have that work out better as a scalar.


Former-commit-id: 1e4050ca5a3c1857dc97e34d74771569f473d32f