Update issue 2998
The check for whether the model was equilibrated during that cycle used to default to the length of the harvest dictionary, so the model was showing a false positive equilibration at the end of the harvest, even if it had just ended due to a constraint on the number of cycles.

Now changed to so that the equilibration point default is -1, so that the equality check for every cycle as it's being brinted to the equilibration point will never match if it didn't actually equilibrate.


Former-commit-id: f94f5b9901dd4c91f9c16ded18ddca45699ae40f