New issue
Adding the current iteration of the pit filling algorithm and the test data to the repository. This issue can be use for commits as we upgrade runtime complexity.

Note: currently uses explicit strings for file locations. This will eventually need to be changed dependant on how we want to do the UI for it.

Former-commit-id: b55501937ec22992065a2ef7a5ebc1eb317db0b4