Update issue 3069
This is a giant push with all the changes for this issue.

I created a prototype that i did not track and then i copied it in and had to make a ton of tweaks.

Now doing validation purely at the raster level and calculating distances using Rich's distance transform function. Not other interpolation should be happening for validation


Former-commit-id: 9aba56520a8334dcd06fc2fdb1239d3002d306d4