Update issue 3081
Making a giant change to handling stresser buffers. Replacing readasarray operations with memory efficient raster util functions. Also editing the functions called on each buffer choice to be memory efficient.

Was able to get one stresser to format correctly in this new method, while a few others are stuck on a distance transform update that Rich just pushed.

I have a lot of debugger statements and intermediate outputs added that will need to be cleaned up.


Former-commit-id: 87b9be12fbd1a7496ff7216544c00a43b6a56b6c