Update issue 3083
Went through each file and added aoi_uri=aoi to all vectorize_datasets calls.
Hopefully this fixes a sizing issue with erroneous data and also makes the model
more complete.

I did not put in documenation for this input yet into the calls / functions it
was added to. That will surely suck


Former-commit-id: a01962f71fdaaaad91eef14d26921f7c41e8208b