Update issue 3190
Fixing an issue now that we do correct nodata casting, except if we plug in a numpy type to set nodata value we get a type error.  So now inside the new data we try to .item() the nodata value just in case it's a numpy type that can be cast appropriately.


Former-commit-id: 726c33b2103cd98bceef2ebcf764ab0cbcae61d3