Fixes issue 3224
raster_utils.aggregate_raster_values_uri is supposed to have placeholder counts for shapefiles that try to aggregate over nodata.  There was a placeholder for pixel_mean setting the value to -9999 if no data were under the polygon but not so for total or hectare_mean.  This caused the hydropower module to raise an exception on water scarcity since it used the total from aggregation, we hadn't detected this issue before because the rest of water scarcity operates on pixel_mean only.  Added -9999 placeholders for total and hectare_mean.


Former-commit-id: d897b8de764eb69495a8d6d698450ef913628a74