Fixes issue 3246
The issue seemed to be gdal.ResizeImage writing bogus '0' values throughout the reinterpolated raster.  We traced this to gdal.GetNoDataValue() returning -128 for an unsigned byte, then setting exactly -128 for the reinterpolated image rather that -128 % 256 == 128 as the reinterpolated image.  So added a wad of code to test for unsigned bytes, shorts, and ints and % them according to their integer size.


Former-commit-id: 0bdb890f14c11b2dbc4257eeaf02795757256b17